Added similar logic to manage bitcoin links
