+++ packages/blockchain-wallet-v4-frontend/src/scenes/Preferences/BitcoinLinkHandling/index.js
@@ -1,35 +1,27 @@
+
 import React from 'react'
-import { FormattedMessage } from 'react-intl'
-import { Button, Text, TextGroup } from 'blockchain-info-components'
-import { SettingComponent, SettingContainer, SettingDescription, SettingHeader, SettingSummary } from 'components/Setting'
+import ui from 'redux-ui'
+import config from 'config'
 
-const BitcoinLinkHandling = (props) => {
-  return (
-    <SettingContainer>
-      <SettingSummary>
-        <SettingHeader>
-          <FormattedMessage id='scenes.preferences.link.title' defaultMessage='Bitcoin link handling' />
-        </SettingHeader>
-        <SettingDescription>
-          <FormattedMessage id='scenes.preferences.link.description' defaultMessage='Enable this to allow your Blockchain Wallet to handle bitcoin payment links in the web browser.' />
-          <FormattedMessage id='scenes.preferences.link.description2' defaultMessage='This will make your experience more convenient when transacting online.' />
-        </SettingDescription>
-      </SettingSummary>
-      <SettingComponent>
-        <Button nature='primary'>
-          <FormattedMessage id='scenes.preferences.link.enable' defaultMessage='Enable' />
-        </Button>
-        <TextGroup inline>
-          <Text size='14px' weight={300} color='error'>
-            <FormattedMessage id='scenes.preferences.link.unknownstatus' defaultMessage="We can't detect whether or not handling of bitcoin links has been enabled." />
-          </Text>
-          <Text size='14px' weight={300} color='error'>
-            <FormattedMessage id='scenes.preferences.link.unknownstatus2' defaultMessage='If it has already been enabled, nothing will happen.' />
-          </Text>
-        </TextGroup>
-      </SettingComponent>
-    </SettingContainer>
-  )
+import BitcoinLinkHandling from './template.js'
+
+class BitcoinLinkHandlingContainer extends React.Component {
+  constructor (props) {
+    super(props)
+    this.handleClick = this.handleClick.bind(this)
+  }
+
+  handleClick () {
+    this.props.updateUI({ warningDisplayed: true })
+    // Register bitcoin links
+    window.navigator.registerProtocolHandler('bitcoin', `${config.ROOT_URL}open/%s`, 'Blockchain')
+  }
+
+  render () {
+    return <BitcoinLinkHandling {...this.props} warningDisplayed={this.props.ui.warningDisplayed} handleClick={this.handleClick}
+    />
+  }
 }
+const enhance = ui({ key: 'BitcoinLinkHandling', state: { warningDisplayed: false } })
 
-export default BitcoinLinkHandling
+export default enhance(BitcoinLinkHandlingContainer)
