+++ packages/blockchain-wallet-v4-frontend/src/scenes/Preferences/Themes/Settings/template.js
@@ -1,18 +1,17 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import styled from 'styled-components'
 import { Field, reduxForm } from 'redux-form'
 
 import { SelectBoxTheme } from 'components/Form'
-import { SettingWrapper } from 'components/Setting'
+import { SettingSelectBoxWrapper } from 'components/Setting'
 
 const Settings = (props) => {
   const { handleClick } = props
 
   return (
-    <SettingWrapper>
+    <SettingSelectBoxWrapper>
       <Field name='theme' component={SelectBoxTheme} callback={handleClick} />
-    </SettingWrapper>
+    </SettingSelectBoxWrapper>
   )
 }
